% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot State Level Accidents (Geo-Plotting)}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character variable representing the State Number reference to use to plot the volume of accidents.}

\item{year}{A character variable representing the year of the file to import.}
}
\value{
A geo-plot of the State Number (\code{state.num}) with the volume of accidents.
}
\description{
This function takes a State Number (\code{state.num}) and a Year (\code{year}) to import a file (based on the year parsed) as a R tibble dataframe.
After filtering the dataframe based on State Number (\code{state.num}), the function outputs a geo-plot of the filtered State Number (\code{state.num}) with the volume of accidents.
}
\details{
The function uses fars_read function (\code{\link{fars_read}}) and make_filename (\code{\link{make_filename}}) to execute the described process.

If State Number (\code{state.num}) doesn't exist on the dataframe an error is returned (i.e. invalid STATE number: 324).
If the filtered dataframe has zero records - the function returns a message 'no accidents to plot'.

Longitudes higher than 900 and Latitudes > than 90 are treated as NAs for plotting purposes.
}
\examples{
\dontrun{
fars_map_state(state.num = "6", year = "2015")
}

}
