% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import Accident Raw Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character value representing the filename to import.}
}
\value{
A tibble dataframe representing the file \code{filename} loaded.
}
\description{
This function takes a filename and load the file into a R tibble dataframe.
If the file path/name is not found or valid, an error message is retrieved.
}
\details{
This function uses and imports readr and dplyr packages form tidyverse package repository.
}
\examples{
fars_read("filename.csv")
}
