% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Import all files to a list of dataframes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character vector of years.}
}
\value{
A list of R tibble dataframes.
}
\description{
This function takes a vector of years and returns a list of R tibble dataframes consisting of the files loaded.
The function loops over the vector of years to load each of the Accident files on the directory whilst manipulating the data on them
to create a new collumn "year" and select only the columns "MONTH" and "year".
}
\details{
The function uses fars_read function (\code{\link{fars_read}}) and make_filename (\code{\link{make_filename}}) to execute the described process.

If an invalid year is retrieved (i.e. if there is no file with the parsed year input), an error is retrieved.

This function uses dplyr package form tidyverse package repository.
}
\examples{
\dontrun{
years <- c("2013","2014","2015")
fars_read_years(years)
}
}
