% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Import all files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character vector of years.}
}
\value{
A R tibble (wide format) dataframe with 'Years' on the rows and 'Months' as columns with the volume of accidents as values.
}
\description{
This function takes a vector of years (\code{years}) to create an R tibble dataframe consisting of the respective files combined (for each year passed in \code{years}).
It then combines together these dataframes into one (using function (\code{\link{fars_read_years}})) to perform aggregations on volumen of accidents per year/month.
}
\details{
The function uses fars_read_years function (\code{\link{fars_read_years}}) to execute the described process.

This function uses dplyr and tidyr packages form tidyverse package repository.
}
\examples{
\dontrun{
years <- c("2013","2014","2015")
fars_summarize_years(years)
}

}
