% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character variable representing the year to pass to the returned filename.}
}
\value{
A file name (character value) using the year passed in the format 'accident_\code{year}.csv.bz2'.
}
\description{
This function takes a (\code{year}) as input and retrieves a character output consisting of a file name.
}
\examples{
make_filename("2015")
}
